input: axp20x-pek: Check for ACPI0011 gpio button device too
The ACPI-6.0 standard defines a standard gpio button device using
the ACPI0011 HID replacing the custom PNP0C40 gpio device used on
many devices with an AXP288 pmic. All DSDTs I've seen for Cherry
Trail devices with an AXP288 pmic define both a PNP0C40 and an
ACPI0011 ACPI device, enabling one or the other depending on whether
the BIOS thinks it is going to boot Android or Windows.

But in the future we may see some Windows only devices only defining
the ACPI0011 device, so lets check for both.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>